+++ src/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: fealonso <fealonso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/17 15:36:21 by fealonso          #+#    #+#             */
-/*   Updated: 2024/09/22 21:11:47 by fealonso         ###   ########.fr       */
+/*   Updated: 2024/09/25 14:52:38 by fealonso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *ptr, int value, size_t num)
 {
 	char	*p;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	p = (char *) ptr;
